pyproject.toml
[project]
name = "langchain-jobrecommendtool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-google-genai>=4.2.1",
    "langchain-text-splitters>=1.1.1",
    "pandas>=2.3.3",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
]
